Fixed Add Book bugs

+++ dotnet/Capstone/DAO/BookDAO.cs
@@ -66,7 +66,7 @@ namespace Capstone.DAO
                         while (reader.Read())
                         {
                             Book book = new Book();
-                            book.ISBN = Convert.ToInt32(reader["isbn"]);
+                            book.ISBN = (int)Convert.ToInt64(reader["isbn"]);
                             book.Author = Convert.ToString(reader["author"]);
                             book.Title = Convert.ToString(reader["title"]);
                             books.Add(book);
